Refactor: Split GetStarted page into components
This commit refactors the GetStarted page into smaller, more focused components for better readability and maintainability. The page is split into Home, HowItWorks, StoryCircles, and SignUp components.
