+++ src/components/get-started/NavItems.tsx
@@ -1,0 +1,50 @@
+
+import React from "react";
+import { Link } from "react-router-dom";
+import { Home, Book, Users, ArrowRight } from "lucide-react";
+
+export type NavItem = {
+  name: string;
+  label: string;
+  path: string;
+  icon: React.ReactNode;
+  ref: React.RefObject<HTMLElement | HTMLDivElement> | null;
+  isButton?: boolean;
+};
+
+export const getNavItems = (
+  homeRef: React.RefObject<HTMLDivElement>,
+  howItWorksRef: React.RefObject<HTMLElement>,
+  storyCirclesRef: React.RefObject<HTMLElement>,
+  signUpRef: React.RefObject<HTMLElement>
+): NavItem[] => [
+  { 
+    name: "home", 
+    label: "Home", 
+    path: "/get-started",
+    icon: <Home className="mr-1 h-4 w-4 sm:h-4 sm:w-4 text-white" />,
+    ref: homeRef
+  },
+  { 
+    name: "how-it-works", 
+    label: "How It Works", 
+    path: "/get-started#how-it-works",
+    icon: <Book className="mr-1 h-4 w-4 sm:h-4 sm:w-4 text-white" />,
+    ref: howItWorksRef
+  },
+  { 
+    name: "join-story-circle", 
+    label: "Join a Story Circle", 
+    path: "/get-started#join-story-circle",
+    icon: <Users className="mr-1 h-4 w-4 sm:h-4 sm:w-4 text-white" />,
+    ref: storyCirclesRef
+  },
+  {
+    name: "sign-up",
+    label: "Sign Up",
+    path: "/get-started#sign-up",
+    icon: <ArrowRight className="mr-1 h-4 w-4 sm:h-4 sm:w-4 text-white" />,
+    ref: signUpRef,
+    isButton: true
+  }
+];
